% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_strava_session.R
\name{generate_strava_session}
\alias{generate_strava_session}
\title{Generate Strava session}
\usage{
generate_strava_session(id, resolution)
}
\arguments{
\item{id}{ID from a given Strava session}

\item{resolution}{Resolution of the data, takes "low", "medium" or "high"}
}
\description{
This function creates a CSV file from a Strava session using Strava's own API.
It requires you have the API set up (see: https://developers.strava.com/).
}
\details{
For using the function, you have to set your specific Strava information as environment variables.
API name must be called "STRAVA_API_NAME", client ID must be called "STRAVA_API_CLIENT_ID,
and the secret key must be called "STRAVA_API_SECRET_KEY".
}
\examples{
session <- generate_strava_session(id = 1292913288, resolution = "high")

}
