% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_garmin_data.R
\name{prepare_garmin_data}
\alias{prepare_garmin_data}
\title{Prepare Garmin session}
\usage{
prepare_garmin_data(dataframe, max_hr = 200, body_weight = 80)
}
\arguments{
\item{dataframe}{Tibble/dataframe}

\item{max_hr}{Maximum heart rate}

\item{body_weight}{Body weight in kilograms}
}
\description{
This function is used for cleaning and preparing a dataframe from a Garmin session.
It can be used directly on a .fit file converted to csv from the Java program RecordToCSV
for Garmin data (https://developer.garmin.com/fit/overview/).
}
\details{
The main functionality is renaming of columns and creating new columns for normalized power and HR.
}
\examples{

session <- prepare_garmin_session(session, max_hr = 200, body_weight = 80)

}
