% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_strava_data.R
\name{prepare_strava_data}
\alias{prepare_strava_data}
\title{Prepare Strava session}
\usage{
prepare_strava_data(dataframe, max_hr = 200, bw = 80)
}
\arguments{
\item{dataframe}{Tibble/dataframe}

\item{max_hr}{Maximum heart rate (BPM)}

\item{bw}{Body weight in kilograms}
}
\description{
This function is the equivalent of prepare_garmin_data for data generated by
generate_strava_session() using the Strava API.
}
\details{
Type "?prepare_garmin_session" in the R console for more information.
}
\examples{
session <- prepare_garmin_session(session, max_hr = 200, body_weight = 80)

}
