% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_labs.R
\name{set_labs}
\alias{set_labs}
\title{Set labs in a ggplot2 object}
\usage{
set_labs(
  pwr_per_kg = NULL,
  hr_per_cent = NULL,
  time_sec = NULL,
  title = NULL,
  subtitle = NULL
)
}
\arguments{
\item{pwr_per_kg}{If you want power in W/kg, set TRUE. If you want absolute power, set FALSE}

\item{hr_per_cent}{If heart rate in per cent of maximum, set TRUE. If you want in BPM, set FALSE}

\item{time_sec}{If you want time in seconds, set TRUE. If you want minutes, set FALSE}

\item{title}{Title}

\item{subtitle}{subtitle}
}
\description{
The function must be passed TRUE/FALSE to time_sec argument, and TRUE/FALSE
to either pwr_per_kg or hr_per_cent
}
