% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot}
\alias{plot}
\title{A ggplot2 based plotting function}
\usage{
plot(
  dataframe,
  x = "x",
  y = "y",
  geom = "line",
  color = NULL,
  fill = NULL,
  alpha = 1,
  linewidth = 0.7,
  xlab = NULL,
  ylab = NULL,
  theme = "theme_light",
  title = NULL,
  subtitle = NULL
)
}
\arguments{
\item{dataframe}{Data frame/tibble}

\item{x}{X axis element}

\item{y}{Y axis element}

\item{geom}{Type of plot, geom object}

\item{color}{Color for the geom object}

\item{fill}{Fill for the geom object}

\item{alpha}{Geom opacity}

\item{linewidth}{Line width (for geom = "line")}

\item{xlab}{X axis title}

\item{ylab}{Y lab title}

\item{theme}{Plot theme}

\item{title}{Plot title}

\item{subtitle}{Plot subtitle}
}
\value{
A ggplot2 object
}
\description{
ggplot2 styled plots for quick and easy plotting. Possible geoms are line, point, area and violin.
The function works in a ggplot workflow, so you can extend or override settings by adding ggplot2 "layers".
}
\examples{
test_data |>
  plot(
  geom = "line",
  x = elapsed_time_min,
  y = heart_rate,
  color = "red",
  alpha = 0.7
  )
}
